requirements.txt
astrapy
openai
langchain
streamlit
tiktoken
pypdf
python_dotenv
unstructured
